requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib
python-jose
python-multipart
pymysql
cryptography
psycopg2
aiosqlite
